Attempting to make the GUI tests actually test the GUIs
