+++ ButtplugKiirooEmulatorGUI/KiirooEmulatorPanel.xaml.cs
@@ -97,6 +97,8 @@ namespace ButtplugKiirooEmulatorGUI
 
         public void StopServer()
         {
+            _bpServer.SendMessage(new StopScanning()).Wait();
+            _bpServer.SendMessage(new StopAllDevices()).Wait(); ;
             _kiirooEmulator.StopServer();
         }
